requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy~=1.26.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow~=10.0.1
pip==23.3.2
psutil~=5.9.6
py-cpuinfo==9.0.0
pynput==1.7.6
pyparsing==3.1.1
pyrealsense2==2.54.2.5684
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
setuptools==69.0.3
six==1.16.0
sympy==1.12
thop==0.1.1.post2209072238
torch~=2.1.0
torchvision==0.17.0
tqdm==4.66.1
typing-extensions==4.9.0
tzdata==2023.4
ultralytics~=8.1.10
urllib3==2.2.0
wheel==0.42.0
zipp==3.17.0

scikit-learn~=1.3.2